requirements.txt
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.8.0
certifi==2024.12.14
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
python-slugify==8.0.4
sniffio==1.3.1
text-unidecode==1.3
typing_extensions==4.12.2
pyyaml>=6.0
